refactor: DB Connector에서 DB 연결을 관리
